+++ src/main/java/com/kurkus/kusinsa/controller/user/CartController.java
@@ -1,4 +1,4 @@
-package com.kurkus.kusinsa.controller;
+package com.kurkus.kusinsa.controller.user;
 
 
 import java.util.List;
