+++ src/main/java/com/kurkus/kusinsa/exception/handler/PointExceptionHandler.java
@@ -3,9 +3,8 @@ package com.kurkus.kusinsa.exception.handler;
 
 import javax.servlet.http.HttpServletRequest;
 
-import com.kurkus.kusinsa.controller.PointController;
+import com.kurkus.kusinsa.controller.user.PointController;
 import com.kurkus.kusinsa.exception.point.PointException;
-import com.kurkus.kusinsa.exception.product.ProductException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
